web_app/service/service_requirements.txt
fastapi
pydantic
pickle5
onnxruntime
numpy
transformers
scikit-learn
uvicorn
requests
beautifulsoup4
mysql-connector-python
python-dotenv
redis